+++ packages/terra-functional-testing/src/commands/utils/index.js
@@ -1,13 +1,17 @@
+const cleanScreenshots = require('./cleanScreenshots');
 const describeTests = require('./describeTests');
 const describeViewports = require('./describeViewports');
+const dispatchCustomEvent = require('./dispatchCustomEvent');
 const getViewports = require('./getViewports');
 const hideInputCaret = require('./hideInputCaret');
 const setApplicationLocale = require('./setApplicationLocale');
 const setViewport = require('./setViewport');
 
 module.exports = {
+  cleanScreenshots,
   describeTests,
   describeViewports,
+  dispatchCustomEvent,
   getViewports,
   hideInputCaret,
   setApplicationLocale,